requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests>=2.25.0
urllib3>=1.26.0
beautifulsoup4>=4.9.0
cloudscraper>=1.2.0
Flask-Migrate==4.0.4
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
SQLAlchemy==2.0.15
uuid==1.30
python-slugify==8.0.1
